requirements.txt
Flask==3.0.0
Flask-Login==0.6.3
Flask-Bcrypt==1.0.1
Flask-WTF==1.2.2
Flask-Mail==0.10.0
Flask-Limiter==4.1.1
Werkzeug==3.0.1
python-dotenv==1.0.0
email-validator==2.3.0
PyPDF2==3.0.1
python-docx==1.1.0
sentence-transformers==2.2.2
transformers==4.36.0
torch==2.1.0
spacy==3.7.2
google-generativeai==0.3.2
pandas==2.1.4
Pillow==10.1.0
gunicorn==21.2.0
